Reorganizar componentes del frontend por responsabilidad
