+++ frontend/src/components/catalog/CatalogFilters.jsx
