+++ frontend/src/components/catalog/CatalogProductCard.jsx
