+++ frontend/src/components/layout/SiteFooter.jsx
@@ -1,4 +1,4 @@
-import Icon from './ui/Icon'
+import Icon from '../common/Icon'
 
 const footerBenefits = [
   {
